pyproject.toml
[tool.poetry]
name = "rpls-back"
version = "0.0.1"
description = ""
authors = ["apierson40 <alexandre.pierson40@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.5"
django-cors-headers = "^3.6.0"
django-rest-framework = "^0.1.0"
django-filter = "^2.4.0"
pandas = "^1.2.0"
requests = "^2.25.1"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
